Delete LICENSE
This repo will contain code from Manning whose license is unclear. It would be inappropriate to imply this is licensed under the GPL without more information. For now, assume Manning maintains copyright on the code and its publishing here on GitHub is implied as a permitted use based on their in-project instructions to do so.